+++ pyinterpolate/variogram/empirical/experimental_variogram.py
@@ -5,7 +5,7 @@ Authors
 -------
 1. Szymon Moliński | @SimonMolinsky
 """
-from typing import Union
+from typing import Union, Dict, Type
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -39,8 +39,8 @@ class DirectionalVariogram:
 
     tolerance : float (in range [0, 1]), default=0.2
         If ``tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
-        the coordinate system and the direction given by y axis and direction parameter. If ``tolerance`` is ``> 0`` then
-        the bin is selected as an elliptical area with major axis pointed in the same direction as the line
+        the coordinate system and the direction given by y axis and direction parameter. If ``tolerance`` is ``> 0``
+        then the bin is selected as an elliptical area with major axis pointed in the same direction as the line
         for 0 tolerance.
 
         * The major axis size == ``step_size``.
@@ -89,13 +89,13 @@ class DirectionalVariogram:
     Methods
     -------
     get()
-        Returns copy of calculated directional variograms.
+        Returns copy of calculated directional variograms or a single variogram in a specific direction.
 
     show()
         Plot all variograms on a single plot.
     """
 
-    def __init__(self, input_array: np.array, step_size: float, max_range: float, weights=None, tolerance=0.2,
+    def __init__(self, input_array: np.array, step_size: float, max_range: float, weights=None, tolerance: float = 0.2,
                  method='t'):
 
         self.ds = input_array
@@ -104,6 +104,7 @@ class DirectionalVariogram:
         self.tolerance = tolerance
         self.weights = weights
         self.method = method
+        self.possible_variograms = ['ISO', 'NS', 'WE', 'NE-SW', 'NW-SE']
 
         self.directions = {
             'NS': 90,
@@ -131,8 +132,30 @@ class DirectionalVariogram:
                                                      method=self.method)
             self.directional_variograms[idx] = variogram
 
-    def get(self):
-        return self.directional_variograms.copy()
+    def get(self, direction=None) -> Union[Dict, Type["ExperimentalVariogram"]]:
+        """
+        Method returns all variograms or a single variogram at a specific direction.
+
+        Parameters
+        ----------
+        direction : str, default = None
+            The direction of variogram from a list of ``possible_variograms`` attribute: "ISO", "NS", "WE", "NE-SW",
+            "NW-SE".
+
+        Returns
+        -------
+        : Union[Dict, Type[ExperimentalVariogram]]
+            The dictionary with variograms for all possible directions, or a single variogram for a specific direction.
+        """
+        if direction is None:
+            return self.directional_variograms.copy()
+        else:
+            if direction in self.possible_variograms:
+                return self.directional_variograms[direction]
+
+            msg = f'Given direction is not possible to retrieve, pass one direction from a possible_variograms: ' \
+                  f'{self.possible_variograms} or leave None to get a dictionary with all possible variograms.'
+            raise KeyError(msg)
 
     def show(self):
         if self.directional_variograms:
@@ -190,8 +213,8 @@ class ExperimentalVariogram:
 
     tolerance : float (in range [0, 1]), default=1
         If ``tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
-        the coordinate system and the direction given by y axis and direction parameter. If ``tolerance`` is ``> 0`` then
-        the bin is selected as an elliptical area with major axis pointed in the same direction as the line
+        the coordinate system and the direction given by y axis and direction parameter. If ``tolerance`` is ``> 0``
+        then the bin is selected as an elliptical area with major axis pointed in the same direction as the line
         for 0 tolerance.
 
         * The major axis size == ``step_size``.
@@ -563,8 +586,8 @@ def build_experimental_variogram(input_array: np.array,
 
     tolerance : float (in range [0, 1]), optional, default=1
         If ``tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
-        the coordinate system and the direction given by y axis and direction parameter. If ``tolerance`` is ``> 0`` then
-        the bin is selected as an elliptical area with major axis pointed in the same direction as the line
+        the coordinate system and the direction given by y axis and direction parameter. If ``tolerance`` is ``> 0``
+        then the bin is selected as an elliptical area with major axis pointed in the same direction as the line
         for 0 tolerance:
 
         * the major axis size == ``step_size``,